curio/ipynb code refactoring and warning feature
